refactor/#17: member 폴더 이름 user로 변경
