+++ src/modules/user/domain/vo/email.vo.ts
