+++ src/modules/user/domain/vo/profile-image-url.vo.ts
