+++ src/modules/user/domain/vo/social-provider.vo.ts
