+++ src/modules/user/user.module.ts
@@ -1,10 +1,10 @@
 import { Module } from "@nestjs/common";
 import { BcryptHasher } from "@src/common/infrastructure/bcrypt-hasher";
-import { UserService } from "@src/modules/member/application/user.service";
-import { UserMapper } from "@src/modules/member/infrastructure/mapper/user.mapper";
-import { PrismaUserRepository } from "@src/modules/member/infrastructure/repositories/prisma-user.repository";
-import { UserController } from "@src/modules/member/presentation/user.controller";
-import { USER_TOKENS } from "@src/modules/member/user.constant";
+import { UserService } from "@src/modules/user/application/user.service";
+import { UserMapper } from "@src/modules/user/infrastructure/mapper/user.mapper";
+import { PrismaUserRepository } from "@src/modules/user/infrastructure/repositories/prisma-user.repository";
+import { UserController } from "@src/modules/user/presentation/user.controller";
+import { USER_TOKENS } from "@src/modules/user/user.constant";
 
 @Module({
   controllers: [UserController],
